% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rsse_ui}
\alias{rsse_ui}
\title{rsse_ui}
\usage{
rsse_ui()
}
\description{
A function to display user friendly interface to do estitamete the sample size for differential expression analysis of RNA-seq data.
}
\examples{
\dontrun{
rsse_ui()
}
}

