% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sample_size}
\alias{sample_size}
\title{sample_size}
\usage{
sample_size(m = 10000, m1 = 100, power = 0.8, f = 0.1, w = 1,
  rho = 2, mu0 = 5, phi_0 = 1, showMessage = F)
}
\arguments{
\item{m}{Total number of genes for testing.}

\item{m1}{Expected number of prognostic genes.}

\item{power}{Power to detecte prognostic genes.}

\item{f}{FDR level}

\item{w}{Ratio of normalization factors between two groups.}

\item{rho}{minimum fold changes for prognostic genes between two groups.}

\item{mu0}{Average read counts for prognostic genes.}

\item{phi_0}{Dispersion for prognostic genes.}

\item{showMessage}{Logical. Display the message in the estimation process.}
}
\description{
A function to estitamete the sample size for differential expression analysis of RNA-seq data.
}
\examples{
\dontrun{#Input initial parameters.
sample_size(m=10000, m1=100, power=0.8, f=0.1, w=1, rho=2, mu0=5, phi_0=1)
}
}

